Create hi.txt
These programs go over the CRUD - (Create, Read, Update, Delete) - operations using PHP and a SQL database. This was inspired by the code walkthrough  in the Web Applications for Everyone course on Coursera. The aim of this project is to hold user "accounts" and allowing the client to perform numerous operations. Furthermore, session data is used to send messages between the pages and erase them after they are displayed.